PLT-3313: Dropping `Arbitrary` subclass relationship in `SemiArbitrary`
