+++ marlowe-spec-test/src/Marlowe/Spec/Core/Semantics.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE BlockArguments #-}
-{-# LANGUAGE TupleSections     #-}
 
 module Marlowe.Spec.Core.Semantics
   ( tests
@@ -8,102 +7,59 @@ module Marlowe.Spec.Core.Semantics
   where
 
 import Arith
-  ( Int (..),
+  ( abs_int,
     less_int,
-    abs_int,
-    integer_of_nat
   )
 import Control.Monad.IO.Class (MonadIO (..))
 import Data.Aeson (ToJSON (..))
 import qualified Data.Aeson as JSON
-import Marlowe.Spec.Core.Arbitrary
-  ( arbitraryTransaction,
-    arbitraryTimeIntervalAfter,
+import Marlowe.Spec.Core.Arbitrary (genEnvironment)
+import Marlowe.Spec.Core.Guarantees
+  ( SemiArbitrary (..),
+    arbitraryTransaction,
     arbitraryValidTransactions,
-    genContract,
-    genEnvironment,
-    genState,
-    genStateWithSize,
-    genValue,
-    genObservation,
-    greater_eq,
-    arbitraryContractWeighted,
-    genGoldenContract
+    genContext
   )
 import Marlowe.Spec.Interpret
   ( InterpretJsonRequest,
     Request (..),
-    Response (..),
+    Response (..)
   )
 import Marlowe.Spec.Reproducible
   ( assertResponse,
     generate,
     generateT,
     reproducibleProperty,
-    reproducibleProperty', ReproducibleTest,
+    reproducibleProperty'
   )
 import Marlowe.Utils (showAsJson)
 import Orderings (Ord (..))
-import PositiveAccounts (validAndPositive_state)
-import QuickCheck.GenT (frequency, suchThat, liftGen, GenT)
+import QuickCheck.GenT (suchThat)
 import Semantics
   ( TransactionOutput (..),
-    TransactionOutputRecord_ext (TransactionOutputRecord_ext),
+    TransactionOutputRecord_ext (..),
     Transaction_ext (..),
     computeTransaction,
-    evalValue,
     evalObservation,
-    isQuiescent,
-    txOutWarnings,
-    txOutPayments,
-    maxTimeContract,
-    inputs,
+    evalValue
   )
 import SemanticsTypes
   ( Contract (..),
     State_ext (..),
     Value (..)
   )
-import SingleInputTransactions (traceListToSingleInput)
-import Test.QuickCheck (cover, withMaxSuccess)
-import Test.QuickCheck.Monadic (assert, monitor, pre, run, stop, PropertyM)
+import Test.QuickCheck (withMaxSuccess)
+import Test.QuickCheck.Monadic (PropertyM, assert, monitor, run)
 import Test.QuickCheck.Property (counterexample)
-import Test.QuickCheck.Arbitrary (Arbitrary (..))
 import Test.Tasty (TestTree, testGroup)
-import Timeout (isClosedAndEmpty)
-import TransactionBound (maxTransactionsInitialState)
-import Debug.Trace
-tests :: InterpretJsonRequest -> TestTree
-tests i = testGroup "Semantics" [ testSemantics i , testGuarantees i ]
 
-testSemantics :: InterpretJsonRequest -> TestTree
-testSemantics i = testGroup "Core"
+tests :: InterpretJsonRequest -> TestTree
+tests i = testGroup "Semantics"
     [ evalValueTest i
     , evalObservationTest i
     , divisionRoundsTowardsZeroTest i
     , computeTransactionTest i
     , computeTransactionForValidTransactionTest i
-    ]
-
--- TODO: Move to its own file Guarantees.hs
--- Property based tests correponding to theorems defined in Isabelle.
-testGuarantees :: InterpretJsonRequest -> TestTree
-testGuarantees i = testGroup "Guarantees"
-    [
-    -- TransactionBound.thy
-    -- TODO: Test skipped as it is currently taking too long
-    -- playTrace_only_accepts_maxTransactionsInitialStateTest i
-    -- SingleInputTransactions.thy
-    -- TODO: Test skipped as it is currently taking too long
-    traceToSingleInputIsEquivalentTest i
-    -- QuiescentResults.thy
-    , computeTransactionIsQuiescentTest i
-    -- TODO: Test skipped as it is currently taking too long
-    -- , playTraceIsQuiescentTest i
-    -- Timeout.thy
-    , timedOutTransaction_closes_contractTest i
-    -- CloseIsSafe.thy
-    , closeIsSafeTest i
     ]
 
 -- The default maxSuccess is 100 and this tests modifies that to 500 as it was empirically found that 10 out of 10 times
@@ -113,9 +69,10 @@ testGuarantees i = testGroup "Guarantees"
 -- are no bugs, only that the selected arbitrary examples didn't find one.
 evalValueTest :: InterpretJsonRequest -> TestTree
 evalValueTest interpret = reproducibleProperty' "Eval Value" (withMaxSuccess 500) do
+    context <- run $ generateT $ genContext interpret
     env <- run $ generate genEnvironment
-    state <- run $ generateT $ genState interpret
-    value <- run $ generateT $ genValue interpret
+    state <- run $ generate $ semiArbitrary context
+    value <- run $ generate $ semiArbitrary context
     let
         req :: Request JSON.Value
         req = EvalValue env state value
@@ -124,9 +81,10 @@ evalValueTest interpret = reproducibleProperty' "Eval Value" (withMaxSuccess 500
 
 evalObservationTest :: InterpretJsonRequest -> TestTree
 evalObservationTest interpret = reproducibleProperty "Eval Observation" do
+    context <- run $ generateT $ genContext interpret
     env <- run $ generate genEnvironment
-    state <- run $ generateT $ genState interpret
-    observation <- run $ generateT $ genObservation interpret
+    state <- run $ generate $ semiArbitrary context
+    observation <- run $ generate $ semiArbitrary context
     let
         req :: Request JSON.Value
         req = EvalObservation env state observation
@@ -135,11 +93,12 @@ evalObservationTest interpret = reproducibleProperty "Eval Observation" do
 
 divisionRoundsTowardsZeroTest :: InterpretJsonRequest -> TestTree
 divisionRoundsTowardsZeroTest interpret = reproducibleProperty "Division rounding"  do
+    context <- run $ generateT $ genContext interpret
     env <- run $ generate genEnvironment
-    state <- run $ generateT $ genState interpret
-    numerator <- run $ generateT $ genValue interpret
-    denominator <- run $ generateT
-        (genValue interpret
+    state <- run $ generate $ semiArbitrary context
+    numerator <- run $ generate $ semiArbitrary context
+    denominator <- run $ generate
+        (semiArbitrary context
           `suchThat` (\d -> abs_int (evalValue env state numerator) `less_int` abs_int (evalValue env state d))
         )
     let
@@ -150,15 +109,17 @@ divisionRoundsTowardsZeroTest interpret = reproducibleProperty "Division roundin
 
 computeTransactionTest :: InterpretJsonRequest -> TestTree
 computeTransactionTest interpret = reproducibleProperty "Calling computeTransaction test" do
-    contract <- run $ generateT $ genContract interpret
-    state <- run $ generateT $ genState interpret
+    context <- run $ generateT $ genContext interpret
+    contract <- run $ generate $ semiArbitrary context
+    state <- run $ generate $ semiArbitrary context
     transaction <- run $ generate $ arbitraryTransaction state contract
     checkComputeTransactionResult interpret contract state transaction
 
 computeTransactionForValidTransactionTest :: InterpretJsonRequest -> TestTree
 computeTransactionForValidTransactionTest interpret = reproducibleProperty "Calling computeTransaction (only valid transactions) test" do
-    contract <- run $ generateT $ genContract interpret `suchThat` (/=Close) -- arbitraryValidTransactions returns [] for the `Close` contract
-    state <- run $ generateT $ genState interpret
+    context <- run $ generateT $ genContext interpret
+    contract <- run $ generate $ semiArbitrary context `suchThat` (/=Close) -- arbitraryValidTransactions returns [] for the `Close` contract
+    state <- run $ generate $ semiArbitrary context
     transactions <- run $ generate $ arbitraryValidTransactions state contract `suchThat` (not . null)
     checkComputeTransactionResult interpret contract state (head transactions)
 
@@ -200,241 +161,3 @@ setEquals :: Eq a => [a] -> [a] -> Bool
 setEquals l1 l2 =
     all (`elem` l2) l1
     && all (`elem` l1) l2
-
--- TransactionBound.thy
---
--- lemma playTrace_only_accepts_maxTransactionsInitialState :
---    "playTrace sl c l = TransactionOutput txOut ⟹
---      length l ≤ maxTransactionsInitialState c"
--- TODO: Test skipped as it is currently taking too long
-_playTrace_only_accepts_maxTransactionsInitialStateTest :: InterpretJsonRequest -> TestTree
-_playTrace_only_accepts_maxTransactionsInitialStateTest interpret = reproducibleProperty "lemma playTrace_only_accepts_maxTransactionsInitialState"  do
-    (contract, State_ext _ _ _ (Int_of_integer startTime) _, transactions) <- run $ generateT $
-        frequency
-          [ (5, genContractStateAndValidTransactions interpret)
-          , (40, genContractStateAndValidTransactions interpret `suchThat` \(_,_,l) -> not (null l))
-          , (55, genContractStateAndValidTransactions interpret `suchThat` \(_,_,l) -> length l > 1)
-          ]
-    let
-        req :: Request JSON.Value
-        req = PlayTrace startTime contract transactions
-    RequestResponse res <- run $ liftIO $ interpret req
-
-    case JSON.fromJSON res of
-      JSON.Success (TransactionOutput TransactionOutputRecord_ext {}) -> do
-        monitor
-          ( counterexample $
-              "Request: " ++ showAsJson req ++ "\n"
-                ++ "Expected maxTransactionsInitialState (" ++ show (maxTransactionsInitialState contract) ++ ")\n"
-                ++ "to be an upper bound for the number of transactions (" ++ show (length transactions) ++ ")")
-        stop $ cover 50.0 (length transactions > 1) "more than one transaction" $ toInteger (length transactions) <= integer_of_nat (maxTransactionsInitialState contract)
-      JSON.Success (TransactionError _ ) -> pre False
-      _ -> fail "JSON parsing failed!"
-
-genContractStateAndValidTransactions :: InterpretJsonRequest -> GenT IO (Contract, State_ext (), [Transaction_ext ()])
-genContractStateAndValidTransactions interpret = do
-    c <- genContract interpret `suchThat` (/=Close)
-    s <- genState interpret
-    t <- liftGen $ arbitraryValidTransactions s c
-    pure (c,s,t)
-
--- SingleInputTransactions.thy
---
--- theorem traceToSingleInputIsEquivalent:
---    "playTrace sn co tral = playTrace sn co (traceListToSingleInput tral)"
--- TODO: Test skipped as it is currently taking too long
-traceToSingleInputIsEquivalentTest :: InterpretJsonRequest -> TestTree
-traceToSingleInputIsEquivalentTest interpret = reproducibleProperty "theorem traceToSingleInputIsEquivalent"  do
-    (contractClass, contract, State_ext _ _ _ (Int_of_integer startTime) _, transactions) <- run $ generateT do
-        (b,c) <- genContract'
-        s <- genState interpret
-        -- This is currently getting stuck on combinations of contract and state that haven't possible actions
-        -- so the only transaction available is [].
-        -- We can put the `suchThat` for all the Gen as before, but is a waste of resources, as a perfectly valid
-        -- contract can be discarded because the state "doesnt fit".
-        -- As a first solution we should try to use a semiArbitrary version for genState. genContract' could
-        -- return as a third value the context for the contract it generates.
-        t <- liftGen $ arbitraryValidTransactions s c
-          `suchThat` \t ->
-              let
-                t' = if (t /= traceListToSingleInput t)
-                  then
-                    trace ("Success: t = " ++ show t ++ "\ntraceListToSingleInput t = " ++ show (traceListToSingleInput t)) t
-                  else
-                    trace ("Fail: t = " ++ show t ++ "\ntraceListToSingleInput t = " ++ show (traceListToSingleInput t) ++ "\ncontract = " ++ showAsJson c) t
-              in
-                t' /= traceListToSingleInput t'
-        pure (b,c,s,t)
-
-    let
-        multipleInputs = PlayTrace startTime contract transactions
-        singletonInput = PlayTrace startTime contract (traceListToSingleInput transactions)
-
-    RequestResponse resMultipleInputs <- run $ liftIO $ interpret multipleInputs
-    RequestResponse resSingletonInput <- run $ liftIO $ interpret singletonInput
-
-    case (JSON.fromJSON resMultipleInputs, JSON.fromJSON resSingletonInput) of
-      (JSON.Success (TransactionOutput _), JSON.Success (TransactionOutput _)) ->
-
-          -- For more than half of the tests the contracts are expected to be arbitrary (i.e. not from golden contracts)
-          stop $ cover 50.0 contractClass "arbitrary contracts" $ resMultipleInputs == resSingletonInput
-
-      (JSON.Success (TransactionError _), JSON.Success _) -> pre False
-      (JSON.Success _ , JSON.Success (TransactionError _)) -> pre False
-      _ -> fail "JSON parsing failed"
-  where
-  genContract' = frequency [(98, (True,) <$> genArbitraryContracts), (2, (False,) <$> genGoldenContract interpret)]
-  genArbitraryContracts = do
-    context <- liftGen arbitrary
-    arbitraryContractWeighted
-      [ (0, 5, 10, 70, 10, 5)    -- A good contract for this test should never start with Close
-      , (5, 10, 10, 60, 10, 5)
-      , (15, 15, 10, 35, 20, 5)
-      , (25, 20, 10, 30, 20, 5)
-      , (25, 20, 10, 30, 20, 5)
-      ]
-      context
-      interpret
-      `suchThat` (\c -> integer_of_nat (maxTransactionsInitialState c) >= 2)
-
-
-isWhen :: Contract -> Bool
-isWhen (When _ _ _) = True
-isWhen _ = False
-
--- QuiescentResults.thy
---
--- theorem computeTransactionIsQuiescent:
---    "validAndPositive_state sta ⟹
---      computeTransaction traIn sta cont = TransactionOutput traOut ⟹
---        isQuiescent (txOutContract traOut) (txOutState traOut)"
--- TODO: Improve test-data quality. Less than 10% of txinput are empty transaction, and manually checking I noticed a lot of timeouted contracts
-computeTransactionIsQuiescentTest :: InterpretJsonRequest -> TestTree
-computeTransactionIsQuiescentTest interpret = reproducibleProperty "theorem computeTransactionIsQuiescent" do
-    contract <- run $ generateT $ genContract interpret `suchThat` (/=Close) -- arbitraryValidTransactions returns [] for the `Close` contract
-    state <- run $ generateT $ genState interpret `suchThat` validAndPositive_state
-    transactions <- run $ generate $ arbitraryValidTransactions state contract `suchThat` (not . null)
-    let
-        transaction = head transactions
-        req :: Request JSON.Value
-        req = ComputeTransaction transaction state contract
-
-    RequestResponse res <- run $ liftIO $ interpret req
-    case JSON.fromJSON res of
-      JSON.Success (TransactionOutput (TransactionOutputRecord_ext _ _ txOutState txOutContract _)) -> do
-        monitor
-          ( counterexample $
-              "Request: " ++ showAsJson req ++ "\n"
-                ++ "Response: " ++ showAsJson res ++ "\n"
-                ++ "Expected reponse to be quiescent")
-        stop $ cover 30.0 (length (inputs transaction) > 0) "Non empty transaction"
-             $ cover 10.0 (isWhen txOutContract) "Output contract is a When statement"
-             $ isQuiescent txOutContract txOutState
-             :: PropertyM ReproducibleTest Bool
-      JSON.Success (TransactionError err ) -> fail $ "Unexpected Transaction Error: " ++ show err
-      _ -> fail "JSON parsing failed!"
-
--- QuiescentResults.thy
---
--- theorem playTraceIsQuiescent:
---    "playTrace sl cont (Cons h t) = TransactionOutput traOut ⟹
---      isQuiescent (txOutContract traOut) (txOutState traOut)"
--- TODO: Test skipped as it is currently taking too long. Not sure why, the data should be the same as for computeTransactionIsQuiescentTest
-_playTraceIsQuiescentTest :: InterpretJsonRequest -> TestTree
-_playTraceIsQuiescentTest interpret = reproducibleProperty "theorem playTraceIsQuiescent" do
-    (contract, State_ext _ _ _ (Int_of_integer startTime) _, transactions) <- run $ generateT $
-        frequency
-          [ (45, genContractStateAndValidTransactions interpret `suchThat` \(_,_,l) -> not (null l))
-          , (55, genContractStateAndValidTransactions interpret `suchThat` \(_,_,l) -> length l > 1)
-          ]
-    let
-        req :: Request JSON.Value
-        req = PlayTrace startTime contract transactions
-    RequestResponse res <- run $ liftIO $ interpret req
-
-    case JSON.fromJSON res of
-      JSON.Success (TransactionOutput (TransactionOutputRecord_ext _ _ txOutState txOutContract _)) -> do
-        monitor
-          ( counterexample $
-              "Request: " ++ showAsJson req ++ "\n"
-                ++ "Response: " ++ showAsJson res ++ "\n"
-                ++ "Expected reponse to be quiescent" )
-        assert $ isQuiescent txOutContract txOutState
-      JSON.Success (TransactionError _ ) -> pre False
-      _ -> fail "JSON parsing failed!"
-
--- Timeout.thy
---
--- theorem timedOutTransaction_closes_contract:
---    "validAndPositive_state sta
---       ⟹  iniTime ≥ minTime sta
---       ⟹  iniTime ≥ maxTimeContract cont
---       ⟹  endTime ≥ iniTime
---       ⟹  accounts sta ≠ [] ∨ cont ≠ Close
---       ⟹  isClosedAndEmpty (computeTransaction ⦇ interval = (iniTime, endTime), inputs = [] ⦈ sta cont)"
-timedOutTransaction_closes_contractTest :: InterpretJsonRequest -> TestTree
-timedOutTransaction_closes_contractTest interpret = reproducibleProperty "theorem timedOutTransaction_closes_contract"  do
-    (contract, state@(State_ext _ _ _ minTime' _)) <- run $ generateT $ (do
-      c <- genContract interpret
-      s <- genState interpret `suchThat` validAndPositive_state
-      pure (c,s)) `suchThat` \(contract, State_ext accounts _ _ _ _) -> not (null accounts) || contract /= Close
-    interval <- run $ generate $ arbitraryTimeIntervalAfter minTime' `suchThat` \(iniTime, endTime) ->
-           (iniTime `greater_eq` minTime')
-        && (iniTime `greater_eq` maxTimeContract contract)
-        && (endTime `greater_eq` iniTime)
-
-    let transaction = Transaction_ext interval [] ()
-        req :: Request JSON.Value
-        req = ComputeTransaction transaction state contract
-
-    RequestResponse res <- run $ liftIO $ interpret req
-
-    case JSON.fromJSON res of
-      JSON.Success txOut@(TransactionOutput (TransactionOutputRecord_ext _ txOutPayment txOutState txOutContract _)) -> do
-        monitor
-          ( counterexample $
-              "Request: " ++ showAsJson req ++ "\n"
-                ++ "txOutContract: " ++ show txOutContract
-                ++ "txOutState: " ++ show txOutState)
-        stop $ cover 100.0 (length (inputs transaction) == 0) "Transaction without inputs"
-             $ cover 10.0 (length txOutPayment > 1) "Timeout contract produces payments"
-             $ cover 70.0 (integer_of_nat (maxTransactionsInitialState contract) >= 1) "At least 1 transaction"
-             $ cover 10.0 (integer_of_nat (maxTransactionsInitialState contract) >= 3) "At least 3 transactions"
-             $ isClosedAndEmpty txOut
-             :: PropertyM ReproducibleTest Bool
-      JSON.Success (TransactionError err ) -> fail $ "Unexpected Transaction Error: " ++ show err
-      _ -> fail "JSON parsing failed!"
-
--- CloseIsSafe.thy
---
--- theorem closeIsSafe :
---    "computeTransaction tra sta Close = TransactionOutput trec ⟹  txOutWarnings trec = []"
-closeIsSafeTest :: InterpretJsonRequest -> TestTree
-closeIsSafeTest interpret = reproducibleProperty "theorem closeIsSafe" do
-    -- NOTE: To satisfy the premise that computeTransaction on a Close results in a succesful TransactionOutput
-    --       we need for the state to have at least one account (Which we guarantee by using the state with size generator)
-    --       This is also shown by the cover test below, which show that all valid transactions necesarly are the empty transaction
-    --       and that it always produce payments. If the accounts were empty then an empty transaction would yield
-    --       the error Useless Transaction, and moreover, there isn't a valid transaction possible
-    --       TODO: Make lemmas about these last two observations.
-    state@(State_ext _ _ _ startTime _) <- run $ generateT $ genStateWithSize ((1, 4), (0, 4), (0, 4)) interpret
-
-    transaction <-  run $ generate $ arbitraryTransaction state Close `suchThat` \(Transaction_ext (_,upper) _ _) -> startTime `less_eq` upper
-    let req :: Request JSON.Value
-        req = ComputeTransaction transaction state Close
-
-    RequestResponse res <- run $ liftIO $ interpret req
-
-    case JSON.fromJSON res of
-      JSON.Success (TransactionOutput trec) -> do
-        monitor
-          ( counterexample $
-              "Request: " ++ showAsJson req ++ "\n"
-                ++ "Expected: no warnings\n"
-                ++ "Actual: " ++ show (txOutWarnings trec))
-        stop $ cover 100.0 (length (inputs transaction) == 0) "Transaction without inputs"
-             $ cover 100.0 (length (txOutPayments trec) > 0) "Close contract produces payments"
-             $ length (txOutWarnings trec) == 0
-             :: PropertyM ReproducibleTest Bool
-      JSON.Success (TransactionError err ) -> fail $ "Unexpected Transaction Error: " ++ show err
-      _ -> fail "JSON parsing failed!"
